Moving a bit of code around to shrink cmd2.py     This includes moving cmd2 decorators into a new file called decorators.py     Moved some constants from cmd2.py to constants.py
