+++ examples/help_categories.py
@@ -7,7 +7,7 @@ A sample application for tagging categories on commands.
 import argparse
 
 import cmd2
-from cmd2.cmd2 import COMMAND_NAME
+from cmd2 import COMMAND_NAME
 
 
 class HelpCategories(cmd2.Cmd):
